refactor: collapse server/grpc sub-module
